+++ dep/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(jwt-cpp)
+add_subdirectory(libbcrypt)
