+++ dep/jwt-cpp/vcpkg/fix-warning.patch
